+++ index.html
@@ -7,14 +7,6 @@
   </head>
 
   <body>
-    <div id="app">
-      <h1>🍿 자판기 🍿</h1>
-      <div class="nav-container">
-        <button id="item-manage-tab" class="nav-button nav-button-clicked">상품 관리</button>
-        <button id="money-charge-tab" class="nav-button">잔돈 충전</button>
-        <button id="item-purchase-tab" class="nav-button">상품 구매</button>
-      </div>
-      <div id="content"></div>
-    </div>
+    <div id="app"></div>
   </body>
 </html>
